record reader options bug fix

+++ models/vgg_vertically_sliced.py
@@ -208,8 +208,8 @@ def build_convolutional_sequential_feature_extractor(inputs, weight_decay, vgg_i
 
     inputs_shape = inputs.get_shape()
     constant_3 = tf.constant(3)
-    horizontal_slice_size = tf.round(tf.divide(inputs_shape[2], constant_3))
-    vertical_slice_size = tf.round(tf.divide(inputs_shape[1], constant_3))
+    horizontal_slice_size = tf.cast(tf.round(tf.divide(inputs_shape[2], constant_3)), tf.int32)
+    vertical_slice_size = tf.cast(tf.round(tf.divide(inputs_shape[1], constant_3)), tf.int32)
     inputs = tf.slice(inputs, begin=[0, vertical_slice_size, 0, 0], size=[-1, -1, horizontal_slice_size * 2, -1])
 
     with tf.contrib.framework.arg_scope([layers.convolution2d],
